requirements.txt
Flask==3.0.2
psycopg==3.1.18
pytest==8.1.1
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
flask-login==0.6.3
colorama==0.4.6
openai==1.59.4
python-dotenv==1.0.1
instaloader==4.14
